address comments, clean backends, pass tests
