+++ geomstats/_backend/pytorch/linalg.py
@@ -12,7 +12,8 @@ def _raise_not_implemented_error(*args, **kwargs):
 eig = _raise_not_implemented_error
 expm = torch.matrix_exp
 logm = _raise_not_implemented_error
-powerm = torch.matrix_power
+inv = torch.inverse
+det = torch.det
 
 
 def cholesky(a):
@@ -23,10 +24,6 @@ def sqrtm(x):
     np_sqrtm = np.vectorize(
         scipy.linalg.sqrtm, signature='(n,m)->(n,m)')(x)
     return torch.from_numpy(np_sqrtm)
-
-
-def inv(*args, **kwargs):
-    return torch.from_numpy(np.linalg.inv(*args, **kwargs))
 
 
 def eigvalsh(a, **kwargs):
@@ -51,14 +48,10 @@ def svd(x, full_matrices=True, compute_uv=True):
     return torch.svd(x, some=not full_matrices, compute_uv=compute_uv)[1]
 
 
-def det(*args, **kwargs):
-    return torch.from_numpy(np.array(np.linalg.det(*args, **kwargs)))
-
-
-def norm(x, ord=2, axis=None):
+def norm(x, ord=None, axis=None):
     if axis is None:
-        return torch.norm(x, p=ord)
-    return torch.norm(x, p=ord, dim=axis)
+        return torch.linalg.norm(x, ord=ord)
+    return torch.linalg.norm(x, ord=ord, dim=axis)
 
 
 def solve_sylvester(a, b, q):
