+++ geomstats/_backend/tensorflow/linalg.py
@@ -66,7 +66,3 @@ def solve_sylvester(a, b, q):
 
 def qr(x, mode='reduced'):
     return tf.linalg.qr(x, full_matrices=(mode == 'complete'))
-
-
-def powerm(x, power):
-    return expm(power * logm(x))
